+++ MagicGradients/Builder/GradientBuilder.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using Xamarin.Forms;
 
-namespace MagicGradients
+namespace MagicGradients.Builder
 {
     public class GradientBuilder : StopsBuilder<GradientBuilder>
     {
