+++ MagicGradients/Builder/LinearGradientBuilder.cs
@@ -1,4 +1,4 @@
-﻿namespace MagicGradients
+﻿namespace MagicGradients.Builder
 {
     public class LinearGradientBuilder : StopsBuilder<LinearGradientBuilder>, IChildBuilder
     {
