+++ MagicGradients/Builder/StopsBuilder.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using Xamarin.Forms;
 
-namespace MagicGradients
+namespace MagicGradients.Builder
 {
     public abstract class StopsBuilder<T>
     {
