+++ MagicGradients/Parser/CssGradientParser.cs
@@ -1,3 +1,4 @@
+using MagicGradients.Builder;
 using MagicGradients.Parser.TokenDefinitions;
 using System.Linq;
