+++ MagicGradients/Parser/TokenDefinitions/ColorNameDefinition.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using MagicGradients.Builder;
+using System;
 using System.Linq;
 using Xamarin.Forms;
