+++ MagicGradients/Parser/TokenDefinitions/LinearGradientDefinition.cs
@@ -1,3 +1,4 @@
+using MagicGradients.Builder;
 using System;
 
 namespace MagicGradients.Parser.TokenDefinitions
