+++ MagicGradients/Parser/TokenDefinitions/RadialGradientDefinition.cs
@@ -1,4 +1,5 @@
-﻿using MagicGradients.Xaml;
+﻿using MagicGradients.Builder;
+using MagicGradients.Xaml;
 using System;
 using Xamarin.Forms;
 using static MagicGradients.RadialGradientFlags;
